refactor(pipeline): separate a module for artifact operations
- [x] separate artifact manipulation functionality from the data pipeline;
